+++ src/tables/expense.rs
@@ -104,6 +104,21 @@ impl Expense {
         .map(|_| {})
     }
 
+    pub async fn db_delete_all(
+        db: Arc<Surreal<Any>>,
+        chat_id: ChatId,
+    ) -> Result<(), surrealdb::Error> {
+        use super::chat::{ID as CHAT_ID, TABLE as CHAT_TB};
+
+        db.query(format!(
+            "DELETE {TABLE}
+             WHERE {CHAT} = ${CHAT_ID}",
+        ))
+        .bind((CHAT_ID, RecordId::from_table_key(CHAT_TB, chat_id.0)))
+        .await
+        .map(|_| {})
+    }
+
     pub async fn db_select(
         db: Arc<Surreal<Any>>,
         chat_id: ChatId,